requirements.txt
# Core dependencies
datasets>=2.14.0
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Retrieval libraries
bm25s>=0.1.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0
gensim>=4.3.0
model2vec>=0.1.0

# API and web
requests>=2.31.0
flask>=2.3.0
streamlit>=1.25.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0